report 댓글기능X 버그 최대한 수정 아마도최종본

+++ src/main/java/com/apartment/apart/domain/community/CommunityRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommunityRepository extends JpaRepository<Community, Long> {
+public interface CommunityRepository extends JpaRepository<Community, Integer> {
     Page<Community> findAll(Pageable pageable);
     Page<Community> findAll(Specification<Community> spec, Pageable pageable);
 }
